feat: move to new command structure
