+++ packages/discord-bot/src/interfaces/Command.ts
@@ -1,0 +1,10 @@
+import {
+  SlashCommandBuilder,
+  SlashCommandSubcommandsOnlyBuilder,
+} from '@discordjs/builders';
+import { CommandInteraction } from 'discord.js';
+
+export interface Command {
+  data: SlashCommandBuilder | SlashCommandSubcommandsOnlyBuilder;
+  execute: (interaction: CommandInteraction) => Promise<void>;
+}
