+++ packages/discord-bot/src/utils/format.ts
@@ -1,0 +1,5 @@
+export const getMsgLink = (
+  guildId: string,
+  channelId: string,
+  msgID: string
+): string => `https://discord.com/channels/${guildId}/${channelId}/${msgID}`;
